[DAPS-14xx] simplify feat
